backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.35
psycopg[binary]
pydantic
pydantic-settings
firebase-admin==6.4.0
python-dotenv==1.0.1
python-multipart==0.0.9
openai>=1.50.0
Pillow
